add docx, shi and malik impl, and sample methods
This does three things, mainly.

First, it implemented shi and malik's recursive bipartitioning algorithm.
This is a different technique from that offered by sklearn, and would have to be
ported to sklearn in order to be useful there.
Otherwise, effectively how it works is to compute the first most meaningful
partition of the spatial/attribute affinity graph. Then, it computes the
first most meaningful partition of each of the subgraphs.
It does this recursively until the number of subgraphs requested has been found.

The nice thing about implementing this method is that we can also do max-P-style
searches, where the number of clusters is left to be unknown, and a floor is specified.
This has been implemented in a way that allows for arbitrary floor weights,
meaning this floor could be in terms of the count of observations in a cluster (default),
or could be in terms of the sum of another attribute.

Second, this adds "sample" methods to spectral clustering, as well as
a subclass of the default agglomerative clustering in Scikit with
sampling methods. This uses a random number generator to give random
weights to the graph, and then partitions using those weights.

Third, this commit implements a significant amount of documentation and quality
of life fixes. Specifically, since the commit adds the recursive bi-partitioning
algorithm, it recognizes that passing np.inf as the number of clusters requested
means that the user must use the recursive bipartitioning algorithm to assign labels.
However, if the number of clusters is still finite, the recursive bipartitioning
can still be used.